+++ packages/cli/src/commands/dev/dev.ts
@@ -127,7 +127,7 @@ const startServer = async (
         if (
           !output.includes('Studio available') &&
           !output.includes('👨‍💻') &&
-          !output.includes('Mastra API running on port')
+          !output.includes('Mastra API running on ')
         ) {
           process.stdout.write(output);
         }
@@ -140,7 +140,7 @@ const startServer = async (
         if (
           !output.includes('Studio available') &&
           !output.includes('👨‍💻') &&
-          !output.includes('Mastra API running on port')
+          !output.includes('Mastra API running on ')
         ) {
           process.stderr.write(output);
         }