+++ client-sdks/react/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @mastra/react-hooks
 
+## 0.0.25-alpha.0
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mastra/client-js@0.17.3-alpha.0
+
+## 0.0.24
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mastra/client-js@0.17.2
+
+## 0.0.24-alpha.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mastra/client-js@0.17.2-alpha.1
+
 ## 0.0.24-alpha.0
 
 ### Patch Changes