+++ packages/_changeset-cli/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @internal/changeset-cli
 
+## 0.0.19
+
 ## 0.0.18
 
 ## 0.0.17