+++ packages/playground-ui/CHANGELOG.md
@@ -1,5 +1,32 @@
 # @mastra/playground-ui
 
+## 7.0.3-alpha.0
+
+### Patch Changes
+
+- Updated dependencies [[`b5581fb`](https://github.com/mastra-ai/mastra/commit/b5581fbcead57443c885f5397d651bf5df20b14a), [`fe5ea7d`](https://github.com/mastra-ai/mastra/commit/fe5ea7d06d56a3ebd3283a25fb45f94aea12d844), [`0ebadd2`](https://github.com/mastra-ai/mastra/commit/0ebadd21d449bd743e5839e875a745afff9f8b0a)]:
+  - @mastra/core@0.24.10-alpha.0
+  - @mastra/client-js@0.17.3-alpha.0
+  - @mastra/react@0.0.25-alpha.0
+
+## 7.0.2
+
+### Patch Changes
+
+- Updated dependencies [[`c56f64b`](https://github.com/mastra-ai/mastra/commit/c56f64b0d7365d744cfe99f54369795917c057b7)]:
+  - @mastra/core@0.24.9
+  - @mastra/client-js@0.17.2
+  - @mastra/react@0.0.24
+
+## 7.0.2-alpha.1
+
+### Patch Changes
+
+- Updated dependencies []:
+  - @mastra/core@0.24.9-alpha.1
+  - @mastra/client-js@0.17.2-alpha.1
+  - @mastra/react@0.0.24-alpha.1
+
 ## 7.0.2-alpha.0
 
 ### Patch Changes